backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
PyJWT==2.6.0
mysqlclient==2.2.0
python-decouple==3.8
Pillow==11.0.0
